move context and contribution to extension
