+++ client/browser/src/shared/components/CodeViewToolbar.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames'
 import * as H from 'history'
 import * as React from 'react'
 import { ActionNavItemsClassProps, ActionsNavItems } from '../../../../shared/src/actions/ActionsNavItems'
-import { ContributionScope } from '../../../../shared/src/api/client/context/context'
+import { ContributionScope } from '../../../../shared/src/api/extension/api/context/context'
 import { ContributableMenu } from '../../../../shared/src/api/protocol'
 import { ExtensionsControllerProps } from '../../../../shared/src/extensions/controller'
 import { PlatformContextProps } from '../../../../shared/src/platform/context'
