+++ client/shared/src/actions/ActionsContainer.tsx
@@ -3,7 +3,7 @@ import React, { useMemo } from 'react'
 import { from } from 'rxjs'
 import { switchMap } from 'rxjs/operators'
 import { wrapRemoteObservable } from '../api/client/api/common'
-import { ContributionScope, Context } from '../api/client/context/context'
+import { ContributionScope, Context } from '../api/extension/api/context/context'
 import { ContributableMenu } from '../api/protocol'
 import { getContributedActionItems } from '../contributions/contributions'
 import { ExtensionsControllerProps } from '../extensions/controller'
