+++ client/shared/src/api/extension/activation.ts
@@ -6,7 +6,7 @@ import { areExtensionsSame } from '../../extensions/extensions'
 import { asError, ErrorLike, isErrorLike } from '../../util/errors'
 import { memoizeObservable } from '../../util/memoizeObservable'
 import { wrapRemoteObservable } from '../client/api/common'
-import { parseContributionExpressions } from '../client/services/contribution'
+import { parseContributionExpressions } from './api/contribution'
 import { MainThreadAPI } from '../contract'
 import { Contributions } from '../protocol'
 import { tryCatchPromise } from '../util'
