+++ client/shared/src/api/extension/api/contribution.test.ts
@@ -1,6 +1,6 @@
 import { ContributableMenu, Contributions, Evaluated } from '../../protocol'
-import { Context } from '../context/context'
-import { parse, parseTemplate } from '../context/expr/evaluator'
+import { Context } from './context/context'
+import { parse, parseTemplate } from './context/expr/evaluator'
 import {
     evaluateContributions,
     filterContributions,
