+++ client/shared/src/components/BadgeAttachment.tsx
@@ -4,7 +4,7 @@ import InformationIcon from 'mdi-react/InfoCircleOutlineIcon'
 import WarningIcon from 'mdi-react/AlertCircleOutlineIcon'
 import ErrorIcon from 'mdi-react/AlertDecagramOutlineIcon'
 import { BadgeAttachmentRenderOptions } from 'sourcegraph'
-import { badgeAttachmentStyleForTheme } from '../api/client/services/decoration'
+import { badgeAttachmentStyleForTheme } from '../api/extension/api/decorations'
 import { LinkOrSpan } from './LinkOrSpan'
 import { isEncodedImage } from '../util/icon'
 import { MdiReactIconComponentType, MdiReactIconProps } from 'mdi-react'
