+++ client/shared/src/hover/actions.ts
@@ -19,7 +19,7 @@ import {
 } from 'rxjs/operators'
 import { ActionItemAction } from '../actions/ActionItem'
 import { wrapRemoteObservable } from '../api/client/api/common'
-import { Context } from '../api/client/context/context'
+import { Context } from '../api/extension/api/context/context'
 import { FlatExtensionHostAPI } from '../api/contract'
 import { WorkspaceRootWithMetadata } from '../api/extension/extensionHostApi'
 import { ContributableMenu, TextDocumentPositionParameters } from '../api/protocol'
