+++ client/web/src/components/diff/FileDiffHunks.tsx
@@ -3,7 +3,7 @@ import * as H from 'history'
 import React, { useCallback, useMemo, useState, useEffect } from 'react'
 import { combineLatest, from, NEVER, Observable, of, ReplaySubject, Subscription } from 'rxjs'
 import { filter, first, switchMap, tap } from 'rxjs/operators'
-import { DecorationMapByLine, groupDecorationsByLine } from '../../../../shared/src/api/client/services/decoration'
+import { DecorationMapByLine, groupDecorationsByLine } from '../../../../shared/src/api/extension/api/decorations'
 import { isDefined, property } from '../../../../shared/src/util/types'
 import { toURIWithPath } from '../../../../shared/src/util/url'
 import { ThemeProps } from '../../../../shared/src/theme'
