+++ client/web/src/repo/blob/Blob.tsx
@@ -6,7 +6,7 @@ import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { BehaviorSubject, combineLatest, EMPTY, from, fromEvent, ReplaySubject, Subject, Subscription } from 'rxjs'
 import { catchError, concatMap, filter, first, map, mapTo, switchMap, tap, withLatestFrom } from 'rxjs/operators'
 import { ActionItemAction } from '../../../../shared/src/actions/ActionItem'
-import { groupDecorationsByLine } from '../../../../shared/src/api/client/services/decoration'
+import { groupDecorationsByLine } from '../../../../shared/src/api/extension/api/decorations'
 import { HoverMerged } from '../../../../shared/src/api/client/types/hover'
 import { ExtensionsControllerProps } from '../../../../shared/src/extensions/controller'
 import { getHoverActions } from '../../../../shared/src/hover/actions'
